fix: update way to handle errors

+++ src/services/UserService/UserService.ts
@@ -1,12 +1,12 @@
 import {User} from 'models/user'
 import {randomUUID} from 'node:crypto'
 import {IRepository} from '../../repository/repository'
-import {InvalidInputError} from '../../utils/errors'
+import {InvalidInputError, NotFoundError} from '../../utils/errors'
 import {userDataValidator} from './userDataValidator'
 import {userRepository} from '../../repository/UserRepository'
 import {IService} from 'services/service'
-import { mapUserDataToDTO } from './mapUserDataToDTO'
-import { validate as isUUID } from 'uuid'
+import {mapUserDataToDTO} from './mapUserDataToDTO'
+import {validate as isUUID} from 'uuid'
 
 class UserService implements IService<User> {
   private repository: IRepository<User>
@@ -40,7 +40,7 @@ class UserService implements IService<User> {
   }
 
   getById = async (id: string): Promise<User | null> => {
-    const isIdValid = isUUID(id) 
+    const isIdValid = isUUID(id)
 
     if (!isIdValid) {
       throw new InvalidInputError()
@@ -62,10 +62,10 @@ class UserService implements IService<User> {
       throw new InvalidInputError()
     }
 
-    const dbUser = await this.getById(id) // check will it throw up to controller level
+    const dbUser = await this.getById(id)
 
     if (!dbUser) {
-      throw new InvalidInputError() // should be 404 not found
+      throw new NotFoundError()
     }
 
     const dataWithValidFields = mapUserDataToDTO(data)
@@ -86,8 +86,14 @@ class UserService implements IService<User> {
       throw new InvalidInputError()
     }
 
+    const dbUser = await this.getById(id)
+
+    if (!dbUser) {
+      throw new NotFoundError()
+    }
+
     const isDeleted = await this.repository.delete(id)
-      
+
     return isDeleted
   }
 }
